fix(domain): remove unused imports and redundant code fix(domain): correct error handling in InvokeJSON and InvokeTxt functions refactor(domain): move JSON and text invocation functions to invoke.go
